@@ -178,7 +178,7 @@ if( preg_match("/^(create_table|alter_table|drop_table|create_index|sql)_?(.*)$/
 						$rename[$m[1]] = $m[2];
 					}
 					//Если было указано изменение добавление колонки
-					if( preg_match("/^\+([^\+\-\s]+):(.+?)$/", $argv[$i], $m) ) {
+					if( preg_match('/^\+([^\+\-\s]+):(.+?)$/', $argv[$i], $m) ) {
 						$add[$m[1]] = $m[2] == "string" ? "varchar(255)" : $m[2];
 					}
 				}
