@@ -1,10 +1,13 @@
 <?php
-/**
- * User: nixx
- * Date: 03.09.14
- * Time: 23:37
- */
-class Boot_Assets {
+namespace Boot;
+
+use Boot;
+use Boot_Exception;
+use Controller_Exception;
+use Exception;
+use Sass;
+
+class Assets {
 
 	/**
 	 * Расширение
@@ -13,10 +16,16 @@ class Boot_Assets {
 	private $ext;
 
 	/**
-	 * Компиляция assests
+	 * Компиляция assets
 	 * @var bool
 	 */
 	private $compile;
+
+	/**
+	 * Сжимает файлы
+	 * @var bool
+	 */
+	private $compress = true;
 
 	/**
 	 * @var bool
@@ -41,6 +50,68 @@ class Boot_Assets {
 		$this->ext = $ext;
 		$this->compile = $compile;
 		$this->debug = $debug;
+	}
+
+	/**
+	 * Вывод ошибки
+	 * @param $message
+	 */
+	public static function error($message) {
+		echo "\x1b[31m{$message}\x1b[0m\r\n";
+		exit(127);
+	}
+
+	/**
+	 * @param $compress
+	 */
+	public function setCompress($compress) {
+		$this->compress = $compress;
+	}
+
+	/**
+	 * Копирует указанные файла в bower.json
+	 */
+	static public function install() {
+		if( file_exists(APPLICATION_ROOT . '/bower.json') ) {
+			$bower = json_decode(file_get_contents(APPLICATION_ROOT . '/bower.json'));
+
+			if( isset($bower->install->path) && isset($bower->install->sources) ) {
+				foreach( $bower->install->sources as $source => $paths ) {
+					foreach( $paths as $path ) {
+						$ext = pathinfo($path, PATHINFO_EXTENSION);
+
+						if( !file_exists(APPLICATION_ROOT . '/' . $path) ) {
+							self::error('Asset ' . $path . ' not found');
+						}
+						if( empty($bower->install->path->$ext) ) {
+							self::error('Asset copy path is not defined: install->path->' . $ext);
+						}
+
+						//Директория размещения ассета
+						$dir = APPLICATION_ROOT . '/public/assets/' . str_replace('bower_components/', '', pathinfo($path, PATHINFO_DIRNAME));
+
+						if( !is_dir($dir) ) {
+							mkdir($dir, 0777, true);
+						}
+						copy(APPLICATION_ROOT . '/' . $path, $dir . '/' . pathinfo($path, PATHINFO_BASENAME));
+					}
+				}
+			}
+		}
+	}
+
+	/**
+	 * @param $link
+	 * @return null|string
+	 * @throws Controller_Exception
+	 */
+	static public function find_path($link) {
+		$dir = APPLICATION_ROOT . '/bower_components';
+
+		if( !file_exists($dir . '/' . $link) ) {
+			throw new Controller_Exception('File not found', 404);
+		}
+		return $dir . '/' . $link;
 	}
 
 	/**
@@ -78,24 +149,24 @@ class Boot_Assets {
 		while( ($row = fgets($f, 4096)) !== false ) {
 
 			//Если попадает под маску
-			if( preg_match("/^" . ($this->ext == "css" ? "\s?\*" : "\/\/") . "=\s(require[^\s]*)\s([^\s\n\r]+)/", $row, $matches) ) {
+			if( preg_match("/^" . ($this->ext == "css" ? "\\s?\\*" : "\\/\\/") . "=\\s(require[^\\s]*)\\s([^\\s\n\r]+)/", $row, $matches) ) {
 
 				//Выполняем функцию
 				switch( $matches[1] ) {
 
 					//Чтение директории
 					case "require_tree":
-						$this->readdir(realpath(APPLICATION_PATH . '/assets/' . $this->ext . '/' . $matches[2]));
+						$this->readdir($matches[2]);
 						break;
 
 					//Чтение директории
 					case "require_directory":
-						$this->readdir(realpath(APPLICATION_PATH . '/assets/' . $this->ext . '/' . $matches[2]), false);
+						$this->readdir($matches[2], false);
 						break;
 
 					//Чтение файла
 					case "require":
-						$this->data .= $this->readfile(realpath(APPLICATION_PATH . '/assets/' . $this->ext . '/' . $matches[2])) . ($this->compile ? PHP_EOL : "");
+						$this->data .= $this->readfile($matches[2]) . ($this->compile ? PHP_EOL : "");
 						break;
 
 					default:
@@ -106,22 +177,65 @@ class Boot_Assets {
 	}
 
 	/**
+	 * Ищет указанный файл
+	 * @param $path
+	 * @return null|string
+	 */
+	public function full_path($path) {
+
+		//Если добавляем из bower
+		if( strpos($path, 'bower_components') === 0 ) {
+			return realpath(APPLICATION_ROOT . '/' . $path);
+		}
+		return realpath(APPLICATION_PATH . '/assets/' . $this->ext . '/' . $path);
+	}
+
+	/**
+	 * Normalize path
+	 *
+	 * @param   string  $path
+	 * @param   string  $separator
+	 * @return  string  normalized path
+	 */
+	public function normalizePath($path, $separator = '\\/') {
+		// Remove any kind of funky unicode whitespace
+		$normalized = preg_replace('#\p{C}+|^\./#u', '', $path);
+
+		// Path remove self referring paths ("/./").
+		$normalized = preg_replace('#/\.(?=/)|^\./|\./$#', '', $normalized);
+
+		// Regex for resolving relative paths
+		$regex = '#\/*[^/\.]+/\.\.#Uu';
+
+		while (preg_match($regex, $normalized)) {
+			$normalized = preg_replace($regex, '', $normalized);
+		}
+
+		if (preg_match('#/\.{2}|\.{2}/#', $normalized)) {
+			throw new \LogicException('Path is outside of the defined root, path: [' . $path . '], resolved: [' . $normalized . ']');
+		}
+
+		return trim($normalized, $separator);
+	}
+
+	/**
 	 * Чтение деректории рукурсивно
 	 * @param $dir
 	 * @param bool $recursive
 	 */
 	public function readdir($dir, $recursive = true) {
-		if( $f = opendir($dir) ) {
+		$realpath = $this->full_path($dir);
+		if( $f = opendir($realpath) ) {
 
 			//Проходим по списку файлов
 			while (false !== ($entry = readdir($f))) {
 				if( !in_array($entry, ['.', '..']) ) {
 
 					//Строим полный путь до файла
-					$path = $dir . '/' . $entry;
+					$path = $this->normalizePath($dir . '/' . $entry);
 
 					//Если это директория
-					if( is_dir($path) ) {
+					if( is_dir($realpath . '/' . $entry) ) {
 
 						//Если нужно проходить по дереву
 						if( $recursive ) {
@@ -138,7 +252,7 @@ class Boot_Assets {
 	/**
 	 * Чтение данных файла
 	 * @param $path
-	 * @throws Exception
+	 * @throws \Exception
 	 * @return string
 	 */
 	public function readfile($path) {
@@ -149,67 +263,77 @@ class Boot_Assets {
 		//Если расширение входим в массив
 		if( in_array($ext, [$this->ext, 'scss']) ) {
 			if( $this->compile ) {
+				$path_new = $this->full_path($path);
+				if( !$path_new ) {
+					echo 'File ' . $path . ' not be found' . PHP_EOL;
+					exit(127);
+				}
+				$path = $path_new;
 
 				//SASS
+				//todo сделать кеширование для development
 				if( $ext == 'scss' ) {
 
 					//Компилируем SASS файл
 					$sass = new Sass();
-					$sass->setStyle(Sass::STYLE_COMPRESSED);
+					$sass->setStyle($this->compress ? Sass::STYLE_COMPRESSED : Sass::STYLE_EXPANDED);
 					$sass->setIncludePath(APPLICATION_ROOT);
-					$sass->setComments(false);
+					$sass->setComments(!$this->compress);
 
-					$filename = pathinfo(APPLICATION_PATH . $path, PATHINFO_FILENAME);
-					file_put_contents('/tmp/' . $filename . '.css', $sass->compileFile($path));
+					$filename = pathinfo($path, PATHINFO_FILENAME);
 
-					//Добавляем префиксы
-					$result = system('postcss --use autoprefixer -o /tmp/' . $filename . '.out.css /tmp/' . $filename . '.css 2>&1', $r);
-					if( $result ) {
-						throw new Boot_Exception($result);
+					//Если компилируем готовые ассеты или в конфиге включен режим префиксов
+					if( $this->compress || !empty(Boot::getInstance()->config->autoprefixer) && Boot::getInstance()->config->autoprefixer ) {
+						file_put_contents('/tmp/' . $filename . '.css', $sass->compileFile($path));
+
+						//Добавляем префиксы
+						$result = system('postcss --use autoprefixer -o /tmp/' . $filename . '.out.css /tmp/' . $filename . '.css 2>&1', $r);
+						if( $result ) {
+							throw new Boot_Exception($result);
+						} else {
+							$css = file_get_contents('/tmp/' . $filename . '.out.css');
+							unlink('/tmp/' . $filename . '.out.css');
+							unlink('/tmp/' . $filename . '.css');
+							return $css;
+						}
 					} else {
-						$css = file_get_contents('/tmp/' . $filename . '.out.css');
-						unlink('/tmp/' . $filename . '.out.css');
-						unlink('/tmp/' . $filename . '.css');
-						return $css;
+						return $sass->compileFile($path);
 					}
-
-					//Ruby SASS
-//					//Компилируем sass
-//					$filename = pathinfo(APPLICATION_PATH . $path, PATHINFO_FILENAME);
-//					$return = system('sass -t compressed --sourcemap=none ' . escapeshellarg($path) . ' ' . escapeshellarg('/tmp/' . $filename . '.css') . ' 2>&1');
-//					if( $return ) {
-//						throw new Boot_Exception('sass error: ' . $return);
-//					}
-//
-//					//Добавляем префиксы
-//					$return = system('postcss --use autoprefixer /tmp/' . $filename . '.css -o /tmp/' . $filename . '_out.css 2>&1');
-//					if( $return ) {
-//						throw new Boot_Exception('autoprefixer error: ' . $return);
-//					}
-//
-//					//Выводим данные
-//					return file_get_contents('/tmp/' . $filename . '_out.css');
 				}
 
 				//Выполняем для обычных css файлов
-				return $this->compress(file_get_contents($path));
-			} else {
-				switch( $this->ext ) {
-
-					case "css":
-						return "<link href='" . str_replace(APPLICATION_PATH, "", preg_replace('/\.scss$/i', '.css', $path)) . "' rel='stylesheet' type='text/css'>" . PHP_EOL;
-						break;
-
-					case "js":
-						return "<script src='" . str_replace(APPLICATION_PATH, "", $path) . "' type=\"text/javascript\"></script>" . PHP_EOL;
-						break;
-
-					default:
-						throw new Exception("Wrong file extension");
+				if( $this->compress ) {
+					return $this->compress(file_get_contents($path));
+				} else {
+					return file_get_contents($path);
 				}
+			} else {
+				return $this->developerTag($path);
 			}
 		}
 		return null;
+	}
+
+	/**
+	 * Сразу вывод
+	 * @param $path
+	 * @return string
+	 * @throws \Exception
+	 */
+	public function developerTag($path) {
+		switch( $this->ext ) {
+
+			case "css":
+				return "<link href='/assets/css/" . preg_replace('/\.scss$/i', '.css', $path) . "' rel='stylesheet' type='text/css'>" . PHP_EOL;
+				break;
+
+			case "js":
+				return "<script src='/assets/js/" . $path . "' type=\"text/javascript\"></script>" . PHP_EOL;
+				break;
+
+			default:
+				throw new Exception("Wrong file extension");
+		}
 	}
 
 	/**
@@ -262,13 +386,13 @@ class Boot_Assets {
 				if( class_exists("Minify_CSS_Compressor", false) == false ) {
 					require_once APPLICATION_ROOT . "/system/boot/extension/Minify_CSS_Compressor.php";
 				}
-				return Minify_CSS_Compressor::process($buffer);
+				return \Minify_CSS_Compressor::process($buffer);
 
 			case "js":
 				if( class_exists("JSMin", false) == false ) {
 					require_once APPLICATION_ROOT . "/system/boot/extension/JSMin.php";
 				}
-				return JSMin::minify($buffer);
+				return \JSMin::minify($buffer);
 		}
 		return "";
 	}
