@@ -1,4 +1,3 @@
 <?
-class DB_Exception extends Exception {
-
-}
+class DB_Exception extends Exception {}
+class ValidateException extends Exception {}
