@@ -34,6 +34,21 @@ class Boot_Params implements Iterator, ArrayAccess {
 	}
 
 	/**
+	 * Получение метода запроса
+	 * @return string
+	 */
+	static public function getMethod() {
+
+		//Получаем основной метод
+		$method = $_SERVER['REQUEST_METHOD'];
+
+		if( strtolower($method) == 'post' && !empty($_POST['_method']) ) {
+			$method = strtoupper($_POST['_method']);
+		}
+		return $method;
+	}
+
+	/**
 	 * Установка разрешенный полей для формы
 	 * @param array $permit
 	 * @return $this
@@ -50,7 +65,8 @@ class Boot_Params implements Iterator, ArrayAccess {
 	public function getValues() {
 		$values = [];
 		if( array_diff(array_keys($this->_params), $this->_permit) ) {
-			Boot::getInstance()->debug('  Unpermitted params: ' . implode(', ', array_diff(array_keys($this->_params), $this->_permit)));
+			$params = array_diff(array_keys($this->_params), $this->_permit);
+			Boot::getInstance()->debug('  Unpermitted params: ' . implode(', ', $params));
 		}
 		foreach( $this->_params as $key => $param ) {
 			if( in_array($key, $this->_permit) ) {
