@@ -1,10 +1,12 @@
 <?php
+use Boot\Core\View;
+
 /**
  * User: Odintsov S.A.
  * Date: 24.05.12
  * Time: 20:42
+ * @deprecated
  */
-
 class Boot_Auth_Lib extends Boot_Abstract_Library {
 
 	/**
@@ -107,6 +109,7 @@ class Boot_Auth_Lib extends Boot_Abstract_Library {
 	 * Http авторизация
 	 * @param $login
 	 * @param $passw
+	 * @deprecated
 	 */
 	static public function httpAuth($login, $passw) {
 		if( isset($_SERVER['PHP_AUTH_USER']) == false || isset($_SERVER['PHP_AUTH_PW']) == false || $_SERVER['PHP_AUTH_USER'] != $login || $_SERVER['PHP_AUTH_PW'] != $passw ) {
@@ -119,7 +122,7 @@ class Boot_Auth_Lib extends Boot_Abstract_Library {
 
 	/**
 	 * Инициализация библиотеки во вьюхе и контроллере
-	 * @param Boot_View|Boot_Layout|Boot_Controller $class
+	 * @param View|Boot_Controller $class
 	 * @return void
 	 */
 	public function init(&$class) {
